Refactor: Move ActiveLabel class to ui.widgets
Move cover art drag and drop implementation to the CoverArtThumbnail widget.
